added duplicate checking and ability to publish new bylineWriter
now checks whether embedded videos, embedded images, article text, the article itself or the bylineWriter record would be a duplicate of any records already in the index and only publishes a new record if not. Also the bylineWriter field can accept either an OIPRef, or a name itself, and if the name hasnt already been registered, it publishes a record to do so. finally, added "public address" to the getWalletBalance endpoint